+++ src/Domain/Admin/FormField/ChecklistField.php
@@ -4,15 +4,15 @@ declare(strict_types=1);
 
 namespace TotalCMS\Domain\Admin\FormField;
 
-class MulticheckboxField extends ChoiceField
+class ChecklistField extends ChoiceField
 {
 	protected string $defaultInputType = 'checkbox';
-	protected string $defaultFieldType = 'multicheckbox';
+	protected string $defaultFieldType = 'checklist';
 
 	protected const INPUT_TYPE             = 'checkbox';
 	protected const OPTION_CLASS           = 'checkbox';
 	protected const LABEL_CLASS            = 'checkbox-label';
-	protected const GROUP_FIELDSET_CLASS   = 'multicheckbox-group';
+	protected const GROUP_FIELDSET_CLASS   = 'checklist-group';
 	protected const GROUP_LEGEND_CLASS     = 'checkbox-group-legend';
 	protected const INPUT_CLASS            = 'checkbox';
 	protected const REQUIRED_ON_CONTAINER  = true;